added cart logical

+++ src/pages/CartItem.jsx
@@ -1,4 +1,5 @@
 import { useDispatch } from 'react-redux';
+import { addItem, minusItem, removeItem } from '../redux/slices/cartSlice';
 
 const CartItem = ({ id, title, type, price, count, imageUrl }) => {
   const dispatch = useDispatch();
@@ -11,17 +12,23 @@ const CartItem = ({ id, title, type, price, count, imageUrl }) => {
     );
   };
   const onClickMinus = () => {
-    dispatch(removeItem(id));
+    dispatch(minusItem(id));
+  };
+
+  const onClickRemove = () => {
+    if (window.confirm('Ты действительно хочешь удалить товар?')){
+      dispatch(removeItem(id));
+    }
   };
 
   return (
     <div class="cart__item">
       <div class="cart__item-img">
-        <img class="pizza-block__image" src={imageUrl} alt="Pizza" />
+        <img class="pizza-block__image" src={imageUrl} alt="Cakes" />
       </div>
       <div class="cart__item-info">
         <h3>{title}</h3>
-        {/* <p>тонкое тесто, 26 см.</p> */}
+        <p>{type}</p>
       </div>
       <div class="cart__item-count">
         <div
@@ -45,7 +52,7 @@ const CartItem = ({ id, title, type, price, count, imageUrl }) => {
             />
           </svg>
         </div>
-        <b>2</b>
+        <b>{count}</b>
         <div
           class="button button--outline button--circle cart__item-count-plus"
           onClick={onClickPlus}
@@ -72,7 +79,10 @@ const CartItem = ({ id, title, type, price, count, imageUrl }) => {
         <b>{price * count} ₽</b>
       </div>
       <div class="cart__item-remove">
-        <div class="button button--outline button--circle">
+        <div
+          class="button button--outline button--circle"
+          onClick={onClickRemove}
+        >
           <svg
             width="10"
             height="10"
